Promote ov018/daPgMthr_c to a single TU (src/actors/d_a_pg_mthr.cpp)
Consolidates 22 legacy per-function sources into one translation unit for
the Mother Penguin actor, ov018 .text 0x02111848-0x0211267c.

The class already carries the cartridge's own RTTI name: 0x02113988 holds
"10daPgMthr_c\0" and 0x0211397c is its type_info, whose base word is
0x0208e390 = _ZTI8dActor_c. No rename was needed.

Destructor defined in the class body, not out of line: the cartridge keeps
D1 (0x02111848) BELOW D0 (0x02111898), and an out-of-line member definition
makes mwccarm emit the group D0-then-D1 plus a homeless D2, which the
whole-range link refuses with "licensed .text functions are not emitted in
ROM address order". In the class body the group comes out ROM-ascending and
no D2 is emitted at all, while the destructor stays the first virtual
declared and so remains the key function.

Matrix4x3 is included via common.h FIRST, deliberately. The type has two
0x30-byte spellings under one guard; func_ov018_02111d28 copies a whole
matrix and the ROM copies it as twelve uniform words (ldm/stm x3), which is
the flat `s32 m[12]` spelling. The nested `{Matrix3x3 r; Vector3 t;}`
spelling scalarizes the last three words into ldr/str pairs and costs an
extra live register.

One genuine misattribution corrected: func_ov018_021122ec was labelled
daObjSm_Lift_c::AfterClsn, but every offset it touches (+0xd4, +0x130,
+0x138, +0x98, +0x374, +0x37c) is a daPgMthr_c offset.

include/decl_common.h: func_ov018_02111bf0 returns void, not int -- ov018
0x02111d18..0x02111d24 pops straight to bx lr with no r0 set. All three
files that referenced it are absorbed by this promotion.

Proof:
  tubuild verify    22/22 MATCH, objisolate clean, reloc destinations clean,
                    all 22 functions in ROM-ascending section order
  tubuild linkcheck SCRATCH-LINK-VERIFIED, ROM identical to the stock build,
                    0 new dsd symbol errors (9 pre-existing, same as baseline)
  rombuild -j16     106/106 exact, 100.000000%, mismatching 0

The ten vague-linkage records the TU emits (six base-class RTTI records,
this class's _ZTI/_ZTS/_ZTV, and Vector3's D1) are licensed in the manifest
as compiler_only_output with their cartridge homes. The _ZTV entry discloses
that romdata_check's comparison is twice limited: it starts at the address
point 0x021139bc, leaving the two preamble words uncompared, and
data_ov018_021139f4 cuts the symbol's extent mid-table, leaving slots 14-30
uncovered.
