+++ attribution.json
@@ -1411,6 +1411,28 @@
     "src/actors/d_a_obj_emm_log.cpp#_ZN13daObjEmmLog_c6RenderEv": "github-actions[bot]",
     "src/actors/d_a_obj_emm_log.cpp#_ZN13daObjEmmLog_c8BehaviorEv": "github-actions[bot]",
     "src/actors/d_a_obj_emm_log.cpp#_ZN13daObjEmmLog_c13InitResourcesEv": "github-actions[bot]",
-    "src/actors/d_a_obj_emm_log.cpp#daObjEmmLog_c_Spawn": "tangosdev"
+    "src/actors/d_a_obj_emm_log.cpp#daObjEmmLog_c_Spawn": "tangosdev",
+    "src/actors/d_a_pg_mthr.cpp#_ZN10daPgMthr_cD1Ev": "andrewboudreau",
+    "src/actors/d_a_pg_mthr.cpp#_ZN10daPgMthr_cD0Ev": "github-actions[bot]",
+    "src/actors/d_a_pg_mthr.cpp#func_ov018_021118fc": "github-actions[bot]",
+    "src/actors/d_a_pg_mthr.cpp#func_ov018_02111968": "github-actions[bot]",
+    "src/actors/d_a_pg_mthr.cpp#func_ov018_02111a48": "github-actions[bot]",
+    "src/actors/d_a_pg_mthr.cpp#func_ov018_02111b3c": "github-actions[bot]",
+    "src/actors/d_a_pg_mthr.cpp#func_ov018_02111bf0": "github-actions[bot]",
+    "src/actors/d_a_pg_mthr.cpp#func_ov018_02111d28": "github-actions[bot]",
+    "src/actors/d_a_pg_mthr.cpp#func_ov018_02111e28": "github-actions[bot]",
+    "src/actors/d_a_pg_mthr.cpp#func_ov018_02111f1c": "github-actions[bot]",
+    "src/actors/d_a_pg_mthr.cpp#func_ov018_02111fac": "github-actions[bot]",
+    "src/actors/d_a_pg_mthr.cpp#func_ov018_021121dc": "github-actions[bot]",
+    "src/actors/d_a_pg_mthr.cpp#func_ov018_02112234": "github-actions[bot]",
+    "src/actors/d_a_pg_mthr.cpp#func_ov018_021122ec": "github-actions[bot]",
+    "src/actors/d_a_pg_mthr.cpp#func_ov018_0211235c": "github-actions[bot]",
+    "src/actors/d_a_pg_mthr.cpp#func_ov018_02112398": "github-actions[bot]",
+    "src/actors/d_a_pg_mthr.cpp#func_ov018_021123d0": "github-actions[bot]",
+    "src/actors/d_a_pg_mthr.cpp#_ZN10daPgMthr_c16CleanupResourcesEv": "github-actions[bot]",
+    "src/actors/d_a_pg_mthr.cpp#_ZN10daPgMthr_c16OnPendingDestroyEv": "andrewboudreau",
+    "src/actors/d_a_pg_mthr.cpp#_ZN10daPgMthr_c6RenderEv": "github-actions[bot]",
+    "src/actors/d_a_pg_mthr.cpp#_ZN10daPgMthr_c8BehaviorEv": "github-actions[bot]",
+    "src/actors/d_a_pg_mthr.cpp#_ZN10daPgMthr_c13InitResourcesEv": "github-actions[bot]"
   }
 }
