+++ config/converted-backslide-exceptions.jsonl
@@ -181,3 +181,9 @@
 {"path": "src/actors/d_a_bg_snwmn.cpp#_ZN11daBgSnwmn_cD1Ev", "reason": "daBgSnwmn_c's D1 (0x02120824) sits BELOW its D0 (0x02120874) in ov072. A real C++ member destructor makes mwccarm emit the D2/D1/D0 triple as one group ordered D0-then-D1, and the whole-range link refuses with 'licensed .text functions are not emitted in ROM address order'. Both variants are therefore carried as '// @symbol' mangled bodies, which costs these two functions the no-mangled-refs and no-raw-offset criteria. Byte-match outranks readability; rombuild 106/106 exact."}
 {"path": "src/actors/d_a_bakubaku.cpp#_ZN12daBakubaku_cD0Ev", "reason": "ov032/daBakubaku_c: D1 (0x021111a0) sits BELOW D0 (0x021111f0) in the cartridge, but mwccarm emits the D2/D1/D0 group D0-first, so a real ~daBakubaku_c() member makes the whole-range link refuse with 'licensed .text functions are not emitted in ROM address order'. Both variants are therefore // @symbol mangled bodies inside extern \"C\" blocks, which costs those two functions the no-mangled-refs and no-raw-offset criteria. The other 21 functions in the TU are unaffected."}
 {"path": "src/actors/d_a_bakubaku.cpp#_ZN12daBakubaku_cD1Ev", "reason": "ov032/daBakubaku_c: D1 (0x021111a0) sits BELOW D0 (0x021111f0) in the cartridge, but mwccarm emits the D2/D1/D0 group D0-first, so a real ~daBakubaku_c() member makes the whole-range link refuse with 'licensed .text functions are not emitted in ROM address order'. Both variants are therefore // @symbol mangled bodies inside extern \"C\" blocks, which costs those two functions the no-mangled-refs and no-raw-offset criteria. The other 21 functions in the TU are unaffected."}
+{"path": "src/actors/daPgMthr_c.cpp#_ZN10daPgMthr_c16CleanupResourcesEv", "reason": "ov018/daPgMthr_c promoted to a single TU (src/actors/d_a_pg_mthr.cpp); the 6 CONVERTED member rows move from the per-symbol src/_ZN10daPgMthr_c*.cpp paths to the promoted file."}
+{"path": "src/actors/daPgMthr_c.cpp#_ZN10daPgMthr_c16OnPendingDestroyEv", "reason": "ov018/daPgMthr_c promoted to a single TU (src/actors/d_a_pg_mthr.cpp); the 6 CONVERTED member rows move from the per-symbol src/_ZN10daPgMthr_c*.cpp paths to the promoted file."}
+{"path": "src/actors/daPgMthr_c.cpp#_ZN10daPgMthr_c6RenderEv", "reason": "ov018/daPgMthr_c promoted to a single TU (src/actors/d_a_pg_mthr.cpp); the 6 CONVERTED member rows move from the per-symbol src/_ZN10daPgMthr_c*.cpp paths to the promoted file."}
+{"path": "src/actors/daPgMthr_c.cpp#_ZN10daPgMthr_c8BehaviorEv", "reason": "ov018/daPgMthr_c promoted to a single TU (src/actors/d_a_pg_mthr.cpp); the 6 CONVERTED member rows move from the per-symbol src/_ZN10daPgMthr_c*.cpp paths to the promoted file."}
+{"path": "src/actors/daPgMthr_c.cpp#_ZN10daPgMthr_cD0Ev", "reason": "ov018/daPgMthr_c promoted to a single TU (src/actors/d_a_pg_mthr.cpp); the 6 CONVERTED member rows move from the per-symbol src/_ZN10daPgMthr_c*.cpp paths to the promoted file."}
+{"path": "src/actors/daPgMthr_c.cpp#_ZN10daPgMthr_cD1Ev", "reason": "ov018/daPgMthr_c promoted to a single TU (src/actors/d_a_pg_mthr.cpp); the 6 CONVERTED member rows move from the per-symbol src/_ZN10daPgMthr_c*.cpp paths to the promoted file."}
