+++ config/converted-baseline.json
@@ -405,12 +405,6 @@
     "src/_ZN10daPgDfdr_c16OnPendingDestroyEv.cpp",
     "src/_ZN10daPgDfdr_cD0Ev.cpp",
     "src/_ZN10daPgDfdr_cD1Ev.cpp",
-    "src/_ZN10daPgMthr_c16CleanupResourcesEv.cpp",
-    "src/_ZN10daPgMthr_c16OnPendingDestroyEv.cpp",
-    "src/_ZN10daPgMthr_c6RenderEv.cpp",
-    "src/_ZN10daPgMthr_c8BehaviorEv.cpp",
-    "src/_ZN10daPgMthr_cD0Ev.cpp",
-    "src/_ZN10daPgMthr_cD1Ev.cpp",
     "src/_ZN11BabyPenguin13OnTurnIntoEggER6Player.cpp",
     "src/_ZN11BabyPenguin13OnYoshiTryEatEv.cpp",
     "src/_ZN11BabyPenguin16CleanupResourcesEv.cpp",
@@ -2442,6 +2436,12 @@
     "src/actors/d_a_obj_wc_obj07.cpp#_ZN15daObjWc_Obj07_c16CleanupResourcesEv",
     "src/actors/d_a_obj_wc_obj07.cpp#_ZN15daObjWc_Obj07_cD0Ev",
     "src/actors/d_a_obj_wc_obj07.cpp#_ZN15daObjWc_Obj07_cD1Ev",
+    "src/actors/d_a_pg_mthr.cpp#_ZN10daPgMthr_c16CleanupResourcesEv",
+    "src/actors/d_a_pg_mthr.cpp#_ZN10daPgMthr_c16OnPendingDestroyEv",
+    "src/actors/d_a_pg_mthr.cpp#_ZN10daPgMthr_c6RenderEv",
+    "src/actors/d_a_pg_mthr.cpp#_ZN10daPgMthr_c8BehaviorEv",
+    "src/actors/d_a_pg_mthr.cpp#_ZN10daPgMthr_cD0Ev",
+    "src/actors/d_a_pg_mthr.cpp#_ZN10daPgMthr_cD1Ev",
     "src/actors/d_a_s_cre.cpp#_ZN8daSCre_c8BehaviorEv",
     "src/actors/d_a_s_cre.cpp#_ZN8daSCre_cD0Ev",
     "src/actors/d_a_s_cre.cpp#_ZN8daSCre_cD1Ev",
