--- src/_ZN10daPgMthr_c16OnPendingDestroyEv.cpp
@@ -1,9 +1,0 @@
-//cpp
-// @symbol _ZN10daPgMthr_c16OnPendingDestroyEv
-/* daPgMthr_c::OnPendingDestroy -- vtable slot 12. The ROM body is empty: the
- * override exists only to occupy the slot. */
-#include "daPgMthr_c.h"
-
-void daPgMthr_c::OnPendingDestroy()
-{
-}
