--- src/_ZN10daPgMthr_cD0Ev.cpp
@@ -1,10 +1,0 @@
-//cpp
-// @symbol _ZN10daPgMthr_cD0Ev
-/* D0 and D1 are two compiler-emitted variants of this one real destructor.
- * Each remains isolated in its own build entry until the original TU is
- * reconstructed physically. */
-#include "daPgMthr_c.h"
-
-daPgMthr_c::~daPgMthr_c()
-{
-}
