--- src/_ZN10daPgMthr_cD1Ev.cpp
@@ -1,7 +1,0 @@
-//cpp
-// @symbol _ZN10daPgMthr_cD1Ev
-#include "daPgMthr_c.h"
-
-daPgMthr_c::~daPgMthr_c()
-{
-}
