--- src/func_ov018_02111d28.cpp
@@ -1,36 +1,0 @@
-//cpp
-// @symbol func_ov018_02111d28
-/* recovered: shared common types */
-#include "common.h"
-
-struct ShadowModel;
-struct dActor_c {
-    void DropShadowRadHeight(ShadowModel &sm, Matrix4x3 &mf, int c, int d, unsigned int e);
-};
-/* Signature deliberately copied from the local declaration above: the
-   ROM name carries by-value class parameters (e.g. Fix12<int>), which
-   mwccarm passes differently at the call site, so declaring the true
-   types breaks the byte match. See notes/mwccarm-codegen.md 6az. */
-extern "C" void _ZN8dActor_c19DropShadowRadHeightER11ShadowModelR9Matrix4x35Fix12IiES5_j(void *, ShadowModel &sm, Matrix4x3 &mf, int c, int d, unsigned int e);
-
-extern "C" void Matrix4x3_FromRotationY(void *m, int angle);
-extern "C" void Matrix4x3_ApplyInPlaceToRotationY(Matrix4x3 *mf, short angY);
-extern "C" void Matrix4x3_ApplyInPlaceToRotationX(Matrix4x3 *mf, short angX);
-extern "C" Matrix4x3 data_020a0e68;
-
-extern "C" void func_ov018_02111d28(dActor_c *self)
-{
-    char *s = (char*)self;
-    Matrix4x3_FromRotationY(s + 0xf0, *(short*)(s + 0x8e));
-    *(int*)(s + 0x114) = *(int*)(s + 0x5c) >> 3;
-    *(int*)(s + 0x118) = *(int*)(s + 0x60) >> 3;
-    *(int*)(s + 0x11c) = *(int*)(s + 0x64) >> 3;
-    _ZN8dActor_c19DropShadowRadHeightER11ShadowModelR9Matrix4x35Fix12IiES5_j(self, *(ShadowModel*)(s + 0x14c), *(Matrix4x3*)(s + 0xf0), 0x140000, 0x50000, 0xf);
-    if (*(short*)(s + 0x380) != 0 || *(short*)(s + 0x382) != 0) {
-        Matrix4x3 *dst = (Matrix4x3*)((*(char**)(s + 0xe8)) + 0xf0);
-        data_020a0e68 = *dst;
-        Matrix4x3_ApplyInPlaceToRotationY(&data_020a0e68, *(short*)(s + 0x382));
-        Matrix4x3_ApplyInPlaceToRotationX(&data_020a0e68, *(short*)(s + 0x380));
-        *(Matrix4x3*)((*(char**)(s + 0xe8)) + 0xf0) = data_020a0e68;
-    }
-}
